rename to singular, use factory for get routes
